% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sG_model.R
\name{simulator_sGaussian}
\alias{simulator_sGaussian}
\title{Simulated example 1 (Gaussian and Cauchy peaks)}
\usage{
simulator_sGaussian(Time, param = c(1, 0.7, 0.01), alpha_norm = c(20,
  60), alpha_cauchy = c(40, 80), mean_global = 0, sigma_a = 5)
}
\description{
Simulated example 1 (Gaussian and Cauchy peaks)
}
\examples{
Time <- seq(0, 200, by = 0.5)
alpha = seq(20, 180, by = 20)
theta = c(1, 0.7, 0.01)
alpha_norm   <- alpha[seq(1,length(alpha), by = 2)]
alpha_cauchy <- alpha[seq(2,length(alpha), by = 2)]
y <- simulator_sGaussian(Time, param = theta, alpha_norm = alpha_norm, alpha_cauchy = alpha_cauchy)
}
