% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sG_model.R
\name{simulator_skew}
\alias{simulator_skew}
\title{Simulated example 2 (Gaussian skewed peaks)}
\usage{
simulator_skew(Time, param = c(1, 0.7, 0.01), alpha_norm = c(20, 40,
  60, 80), mean_global = 0, sigma_a = 5)
}
\description{
Simulated example 2 (Gaussian skewed peaks)
}
\examples{
Time <- seq(0, 200, by = 0.5)
alpha = seq(20, 180, by = 20)
theta = c(1, 0.7, 1)
y <- simulator_skew(Time, alpha_norm = alpha, param = theta)
}
